Rename aux_tracers to specialized_tracers
Update all CMake references and include paths to use specialized_tracers
instead of aux_tracers for water tracer and isotope processes.
