+++ components/eamxx/src/physics/CMakeLists.txt
@@ -20,4 +20,4 @@ if (SCREAM_ENABLE_MAM)
   add_subdirectory(mam)
 endif()
 add_subdirectory(gw)
-add_subdirectory(aux_tracers)
+add_subdirectory(specialized_tracers)
